move lma_intercept_rhi to lmalib
